Restructure the HTML parser into its own module
Restructure and separate functionalities/definitions for neatness.

Relevant Issue: #2
