+++ htmlparser/htmlparser.py
@@ -1,0 +1,13 @@
+# Reference: https://html.spec.whatwg.org/
+
+
+def parse_html(html: str) -> None:
+    """Parse HTML string.
+
+    Args:
+        html (list[str]): The full html string
+
+    Returns:
+        None
+
+    """
